[TASK] Update database schema as first and last update wizard
Introduces two new upgrade wizards in the Install tool.

The first wizard - added as first step of the upgrade wizards - adds
tables, fields and keys to comply to the database schema. When this is
necessary no other wizards can be executed until these are created.

The second wizard - added as last step of the upgrade wizards - changes
tables, fields and keys to comply to the database schema. When other
upgrade wizards are available, this one is not available to make sure
they have all necessary fields.

In order to make sure they are added as first and last step they are
added in UpdateWizard instead of ext_localconf.php.

The former "Final step" is now optional and has been renamed to "Hint".
The buttons to start the update wizards from the list have been renamed
from "Next" to "Execute".

Resolves: #53890
Releases: 6.2
Change-Id: I866b558df3325acca3122bbd4e0c2285447fcdf3
Reviewed-on: https://review.typo3.org/27240
Reviewed-by: Markus Klein
Tested-by: Markus Klein