[TASK] Improve revExplode performance
``GeneralUtility::revExplode()`` is implemented in a very
performance-intensive way: it always does a complete cycle with multiple
array operations, be it necessary or not.

To make the method more performant, it is split into different cases.
Currently, the only used value for the ``$count`` in the core is 2, but
we have to implement all others, too, of course. The performance
increase for $count=2 is greater than 50%; the others will at maximum be
as bad as the old implementation, plus a very tiny overhead for checking
the value of ``$count``.

Change-Id: I79d9a87b790935415c636ee87e6a6db952e0baa4
Resolves: #55564
Releases: 6.2
Reviewed-on: https://review.typo3.org/27255
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring