[TASK] Speed up page ID resolving in Extbase BE
Extbase does some simple calls to determine the
page ID of the current request, and does not cache it.

This costs some mySQL calls, which are unneeded
once the current page ID is found. The page ID is
now stored in the ConfigurationManager to retrieve
with a simple runtime cache.

Resolves: #55566
Resolves: #57125
Releases: 6.2
Change-Id: Id5fe4a25591c77ce8a81cefa018a04e803ba7e8a
Reviewed-on: https://review.typo3.org/27243
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind