[BUGFIX] Extbase registers unneeded code for plugins
Every time when an extension is registering an extbase
plugin (in ext_localconf.php), the call triggers some
additional TypoScript code, which is basically empty
and does some superfluous code.

The attached patch removes the code inside the
registration process in order to remove the memory
footprint and have a small impact on performance
when bootstrapping TYPO3.

On a frontend script (eID), this makes approx.
2% of the bootstrap process until TSFE kicks in.

Releases: 6.2
Resolves: #55569
Change-Id: I39b810d06be57f436a68afaa717d5ec214b8797b
Reviewed-on: https://review.typo3.org/27248
Reviewed-by: Felix Oertel
Tested-by: Felix Oertel