[BUGFIX] Framework Packages aren't sorted properly
In the package dependency resolver, the method
buildDependencyGraph builds the dependency graph
first for the framework packages and afterwards
for all left packages.

The filtering if a package is a framework package
happens by comparing the package path with PATH_typo3
and a sysext folder contant. But as PATH_typo3 is
absolute and the package path from the package states
configuration is relative to PATH_site, this filter
does not work.

The patch changes the path to a relativ one.

Fixes: #55623
Releases: 6.2
Change-Id: I2f57dee433054463ebfd17aab5f76b19986c2747
Reviewed-on: https://review.typo3.org/27300
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring