[TASK] Remove rollover functionality from TMENU
The TMENU code has "RO" functionality in its class.
This code might be used from TMENU_LAYERS, but
not from the core itself. The core should remove the
according places. If extensions need this functionality
they could implement their own menu functionality.

The patch itself just removes the functionality of
rollovers, which is currently not documented nor used.

The idea of the content object with its extProc classes
is outdated and should be replaced by hooks and/or
signal slots.

As all methods that are removed, are private, they can
IMHO be removed directly.

In a second patch, there will be a hook interface
in order to manipulate the items.

Releases: 6.2
Resolves: #55751
Change-Id: Ibd5862124a6f8bc9f06254234cfbf179893c27cc
Reviewed-on: https://review.typo3.org/27396
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer