[TASK] Switch info and func module to mod dispatcher
BE modules info and func now uses mod.php.

Changing the entry script for func and info module
will cause integrated submodules to fail in
case they use the old entry point index.php
with a relative path somewhere in funcMenu or in
other links or actions.

A compatibility layer is introduced in
typo3/index.php to catch those old
links and redirect to mod.php with proper parameters.

Change-Id: I83d31b6004ae3556dd8bf7c5c47b0fea39aaad55
Resolves: #55809
Resolves: #55672
Resolves: #55670
Releases: 6.2
Reviewed-on: https://review.typo3.org/27508
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn