[BUGFIX] Upgrade Wizard fails with exception
The package manager throws an exception "core" package
already being registered. This is caused by the signal
which instantiates the PackageManager while in
install tool the actual class is FailsafePackageManager

Solution is to not register the slot in install tool
as the failsafe package manager already takes care
of scanning for available packages.

Releases: 6.2
Resolves: #55969
Change-Id: I34b33e4ed947b532c24265e16dcc52a7aac4487e
Reviewed-on: https://review.typo3.org/27616
Reviewed-by: Wouter Wolters
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel