[BUGFIX] getLabelsFromItemsList() returns no value
Method getLabelsFromItemsList() is expected to return an empty list
when the column is populating items from a foreign_table configuration
option. Method getProcessedValue() is explicitly checking this to
retrieve the label from the corresponding foreign record if it could not
be resolved from the list of static items.

This reverts commits 5dd32b83e7085ed44713c9b97d5c2c47d820765e and
9596d4da3ac8110a76cb3eadabb9307b67971793 and moves the business logic
into method getProcessedValue().

Releases: 6.2, 6.1
Fixes: #54131
Change-Id: I15dcf0c4bcb76ecc85de8e3202b1376a24981b06
Reviewed-on: https://review.typo3.org/27689
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel