[TASK] Return ProcessedFile instead of File if possible
You can query the FAL-API for files by their filename. If the file resides
in the processing-folder of a storage it is a valid file to the storage.
Therefore creating a file-object and indexing it would be a valid thing.

Since this is not intended additionally to the storage the ResourceFactory
must know about that, too. This change implements this awareness, so there
is no way to create a FileObject for files which reside within the
processing folder.
In 6.0 and 6.1 there has been a dirty workaround for that by marking those
files as not indexable which was removed silently when introducing the new
indexer.

Releases: 6.2
Resolves: #54234
Change-Id: I274c47717560a6efc175fa94aacc4770a578fd3f
Reviewed-on: https://review.typo3.org/27753
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Markus Klein
Tested-by: Markus Klein