[FEATURE] Allow Storages outside the webroot
Storages can be marked as not public or be outside of webroot.
There are no direct links to the files posible.

This change adds a eID implementation to create urls to these files.
It also implements a hook in het eID dispatcher that makes is possible
to add custom security/access checks when delivering files.

Releases: 6.2
Resolves: #36420
Change-Id: Ie7fc9c8a674a3200ea2b29d1adc17327d10bd6e5
Reviewed-on: https://review.typo3.org/27760
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter