[BUGFIX] Fix JS concat if first file is forced on top
If a javascript file is forceOnTop the script uses array_unshift()
to put the file at the beginning of the section array. If this is
the first file that got processed, the array does not exist yet,
so the array_unshift function returns NULL and the file
will not added to the array.

The fix properly initalizes the section sub-array.

Resolves: #56242
Releases: 6.2, 6.1
Change-Id: Iadf2b6f2f512929ff5e8598af9972c5337949341
Reviewed-on: https://review.typo3.org/27818
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack