[BUGFIX] New filename does not get sanitized
When adding a new file through the ResourceStorage there
is a check to see if the file already exists. But this check
does not sanitize the target filename, so it could happen that
you get a false positive because when the file really is added
to the file system the target filename is sanitized.

This patch sanitizes the file name before the fileExists check.

Releases: 6.2, 6.1
Resolves: #55299
Change-Id: I519220040448b08883146caf463ed58544a18453
Reviewed-on: https://review.typo3.org/27806
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Xavier Perseguers
Reviewed-by: Wouter Wolters
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel