[CLEANUP] Remove outdated $CLIENT[FORMSTYLE]
The functionality for having to deal with
browsers that did not support style="width: XYpx"
is not needed anymore, as any browser supported
in the TYPO3 universe can deal with that (even IE6).

This patch cleans up all the areas that dealt with
that part. Also, one dummy function in
FormEngine.php has been removed completely,
another in DocumentTemplate.php has been
deprecated.

Releases: 6.2
Resolves: #56254
Change-Id: I6f47a5f7cabfd340088c242f1ee15b83c7cba0fe
Reviewed-on: https://review.typo3.org/27824
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack