[BUGFIX] Admin panel shows wrong cache information
When labels were extracted to translation files, a error was made
that changed the meaning of the admin panel entry. But the code
was not modified to reflect the change of the meaning. Thus the
admin panel shows that the page is cached when it is actually not
cached and vice versa. Additionally, it shows 0 and 1 for the
"Page cached" entry. This should be "yes" or "no", not numbers.

Change-Id: I2519593e7da08b71ccdd5d7dc3dd6ced7ecac3ac
Resolves: #56265
Releases: 6.3, 6.2
Reviewed-on: http://review.typo3.org/31868
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Reviewed-by: Pascal Dürsteler <pascal.duersteler@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>