[BUGFIX] Fix xml2array values for attributes type NULL
Method array2xml converts elements with value NULL to
nodes with attribute type "NULL" and empty value, e.g.:
<node index="myIndex" type="NULL"></node>
This is not correct transformed back in xml2array and results
in an empty string as value.
This in turn results in wrong imported data using impexp with
xml format.

Resolves: #55082
Releases: 6.2
Change-Id: Icfea2f2b7524c44711866c3f75f06331a1958183
Reviewed-on: https://review.typo3.org/27845
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Markus Klein
Tested-by: Markus Klein