[TASK] Form sysext entry script cleanup
The form sysext contains a wizard to provide help for form creation.
This script used the init.php entry script. This is no longer wanted,
but the API can be used to add wizards.

To do so, some files were moved around and follow the same schema
as shown in openid sysext.

Change-Id: I6b6c41de6c271c56404151b1ec95f45e6f7786af
Resolves: #55669
Releases: 6.2
Reviewed-on: https://review.typo3.org/27347
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Markus Klein
Tested-by: Markus Klein