[TASK] Redirect to distributions after first login
For the users to select a distribution, the installation should
redirect directly to the extension manager, after the first
login of the admin, who set up the site.

Then, the list of TER extensions is loaded initially and all
distributions are shown.

For this to happen, the install tool needs to add another option
that sets a UC flag in the just created admin user.
This UC flag is evaluated when the start module is to be shown.

Also, it adds some more logic when redirecting to the start module
to allow to directly jump to a modfunc/action.

Additionally, the distribution action should initially load all
extensions from TER on first hit.

Releases: 6.2
Resolves: #56321
Change-Id: I1e86b5804011e84f7936514e4b88c4a257905e56
Reviewed-on: https://review.typo3.org/27866
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn