[BUGFIX] Respect table mapping on caching
The method analyzeFields() of DBAL's DatabaseConnection caches the
table structure. The cache always uses the real table name, ignoring
any mapping, which results in a RuntimeException:

"Could not update BLOB >>>> no WHERE clause found!"

The method now checks for a possible mapping and sets the alias
as table name.

Resolves: #56349
Releases: 6.2
Change-Id: I9599cebc8604103c3749bc142cd44b813f018a72
Reviewed-on: https://review.typo3.org/27874
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers