[BUGFIX] Unnatural processing order in IRRE tests
The IRRE tests for creating and modifying records first process
offer, hotel then content which is the unnatural order compared
to the processing delivered by the FormEngine (t3lib_TCEforms).
To avoid incorrect testing behaviour and invalid testing
results, the order is changed to content, hotel and offer
(parent to child).

Resolves: #56374
Releases: 6.2
Change-Id: Ib14fc2d50b493e4be982faeba77401b320868639
Reviewed-on: https://review.typo3.org/27887
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert