[!!!][SECURITY] Allow first install only with FIRST_INSTALL file
It was previously possible to access the install tool
by using a specially crafted URL, which caused
the install tool enable file check to fail.

As there was no easy solution to solve this issue,
we now introduce the need to create a file on first
install.

So in the installation directory the following must
be present:

d typo3
f index.php
f FIRST_INSTALL

After the installation the file will be removed.

Resolves: #55387
Releases: 6.2
Change-Id: I583581f18b939ba032950451bab17ac20131683b
Reviewed-on: https://review.typo3.org/28612
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel