[BUGFIX] Allow search for FAL related DB records
The backend search is currently not
working for file collections etc as there
is no searchFields string provided.

The patch adds them, and also
adds it to non-visible records like
sys_file and sys_file_records as they
might be used in a different
visualization when having a filesearch
service.

The patch is easily testable if you
take a file collection and name it
"my collection". Searching in the list
module on that page for "collection"
does not show anything without
the patch.

see
https://review.typo3.org/#/c/16725/9

Releases: 6.2
Resolves: #56410
Change-Id: I0e99b3b291f085b81560e8f823d3e258a8645fc0
Reviewed-on: https://review.typo3.org/27928
Reviewed-by: Tom Ruether
Tested-by: Tom Ruether
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters