[BUGFIX] Setting capabilities to readonly show always drag+drop dialog
If a driver is readonly, the upload button in the file list disappears,
but the DragUploader is still activated - on the whole page.
Clicking anywhere on the list page causes an upload file selection to
pop up. Additionally the new button is shown but has no functionality.
The patch removes DragUploader and superfluous button.

Change-Id: I3f6c2e932d9f66feb6590f08229ddaaad06e688e
Resolves: #56443
Releases: 6.2
Reviewed-on: https://review.typo3.org/27946
Reviewed-by: Nicole Cordes
Tested-by: Nicole Cordes
Reviewed-by: Christian Weiske
Tested-by: Christian Weiske
Reviewed-by: Rico Sonntag
Tested-by: Rico Sonntag
Reviewed-by: Frans Saris
Tested-by: Frans Saris