[BUGFIX] Always fetch related objects if relation set by child
If the DataMapper shall map a 1:N relation for
a property of a domain object and the element
type of this property is another domain object
and not a storage type, the data mapper must
always fetch the related elements from persistence
and cannot rely on the persistence session lookup.

This is the case because the relation to the parent
is set on the child side and not the other way around.
The column in the parent row is useless in that case
and does not hold the id of the child.

We can also not rely on the persistence session
because we do not know the identifier of the child
until we fetched it.

Resolves: #56442
Releases: 6.2, 6.1
Change-Id: Icc3ebf9b825f6380691c60839621f01ca9875e4e
Reviewed-on: https://review.typo3.org/27949
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs