[!!!][SECURITY] Remove old backend_layout wizard
Keeping the old wizard script would not solve
the CSRF attack vector as they could still
be referenced in this kind of attack.

Because of that, we remove it now.

This change provides a backwards compatibility layer.

It will however break code which link to the
old scripts directly in other places.

Resolves: #56625
Releases: 6.2
Change-Id: I07577dca0e16cf095e114799ace4a6e344ad5aa3
Reviewed-on: https://review.typo3.org/28121
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel