[BUGFIX] Fix alignment for button "copy default content elements"
In the page module in the "languages" view, the alignment of the button
"copy default content elements" is off. The button needs some padding on
the left side.

This patch fixes this by adding a div which gets the padding. It also
replaces the br's in front of this button with CSS for spacing.

Change-Id: Ic21097e8db078f2a2612c89d9496bf20fa5c0585
Resolves: #56636
Releases: 6.2
Reviewed-on: https://review.typo3.org/28214
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Sebastian Fischer
Tested-by: Sebastian Fischer
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters