[BUGFIX] Inconsistent visual appearance Install Tool
The install tool function System Environment is not according to the
style sheet of the install tool. This is due to the fact phpinfo() is
called on this page which comes with its own CSS styles. These styles
are influencing the visual appearence of the rest of the page.

This patch adds a new utility class to cleanup the output of phpinfo()
before assigning it to the view. It also comes with our own CSS for
phpinfo().

Change-Id: Ib0b087cae4bf5db723b0e9c3481619c984ef674f
Resolves: #55116
Releases: 6.2
Reviewed-on: https://review.typo3.org/28225
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Frans Saris
Tested-by: Frans Saris