[TASK] Reduce record data in functional test failure message
If records are found in a database table that are not part of
the assertion data-set, currently the whole record array from
the database is shown as failure.
The assertion has information about the fields that should have
been asserted. To keep a better overview, the failure message
is reduced now to exactly those fields.

Resolves: #56803
Releases: 6.2
Change-Id: Ib8f2afd0d464e8a4b850929277508770ff56eb91
Reviewed-on: https://review.typo3.org/28285
Reviewed-by: Wouter Wolters
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader