[BUGFIX] Layout category tree is broken
There are several issues with the layout of the category tree.

When hovering the whole element, the categories (text) get underlined.
The categories text is not aligned properly with the icon and checkbox.
The glow on a checkbox with focus is bigger than the checkbox itself.

This patch fixes these issues by adding CSS for the tree when used in
TCEforms.

Change-Id: Id9faaf906c786a95246d12f15821372648fc44b6
Resolves: #56752
Releases: 6.2
Reviewed-on: https://review.typo3.org/28250
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Francois Suter
Tested-by: Francois Suter