[BUGFIX] Fix thumbnail display for IRRE records
The system will only generate a thumbnail for IRRE records if the
related file object is an image file.

Additionally the UID of the file reference will be used as alternative
text and as title for the thumbnail so that it behaves like the icon
that is used for non-image files.

Resolves: #56274
Releases: 6.2
Change-Id: Ibe6530368434c384d5b7cca93e81b656c8c56383
Reviewed-on: https://review.typo3.org/27844
Reviewed-by: Frans Saris
Tested-by: Frans Saris
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind