[TASK] Remove unmaintainable testclass
The test class contains only one test, but to execute this,
it depends on many classes which are not all properly mocked and a
nightmare to maintain. Additionally the test has nothing
to do with the execution of tasks, so subject scope is not met.

To keep test maintenance a doable task this test class in removed.

Change-Id: I97ac02040cb6dc711dcc300fb84c15c6785736e2
Resolves: #56924
Relates: #51436
Releases: 6.2
Reviewed-on: https://review.typo3.org/28401
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind