[TASK] inline BEUserAuth::getCookieName
During bootstrapping in index_ts.php there is a check
for an existing be_typo_user cookie which enables
TimeTracking. The simple static getter triggers the loading
and parsing of BackendUserAuthentication and
AbstractUserAuthentication which are both pretty fat classes
and have an impact upon eID calls and frontend calls.

This patch inlines the cookie name retrieval and defers thus
the loading of the fat classes.

Resolves: #56316
Releases: 6.2
Change-Id: I5ce5cf92095dcb8820fb0e548e16e38a99f4f12e
Reviewed-on: https://review.typo3.org/28418
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter