[TASK] Tests: Skip some FAL tests
While the FAL tests were already streamlined and simplified a lot,
the ResourceStorageTest class is still a mixture of functional and
unit test that involves heavy set up code and still does not mock
dependencies properly.
This test class needs a rewrite and a clean separation of units
together with a number of working functional tests to be of any use.
For now, all tests that rely on initialized database are skipped and
marked as TODO.

Change-Id: I3191c809c50f4f19ab91b9a5b5f2f3ebc3fa47ff
Resolves: #56940
Related: #51436
Releases: 6.2
Reviewed-on: https://review.typo3.org/28424
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters