[TASK] Tests: Remove extbase BaseTestCase
Lots of unit tests did inherit from extbase BaseTestCase. Meanwhile
all code of this abstract was merged to ext:core UnitTestCase and
with the refactoring of the core unit test boostrap the objectManager
that is still in this class does not work anymore.
The class is now removed, but since this is probably used in some
third party extensions, it is set as class alias to ext:core
UnitTestCase and marked as deprecated in a lecagy file, so IDEs
hint developers about this deprecation.

Change-Id: I31fb337e065aed7f03351cb04e13c68a98c93a34
Resolves: #56963
Related: #51436
Releases: 6.2
Reviewed-on: https://review.typo3.org/28445
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters