[BUGFIX] Exception in Locker if typo3temp does not exist
Follow-up to 2268f64c18b607ba39d06b1580afb0ccd786d6b1

If typo3temp does not exist, the Locker should throw an exception.
The ClassLoader needs to catch this exception and acts upon
depending on whether we are in the Installer or not.

When in the Installer, locking has to be skipped completely due to
missing typo3temp directory.

Resolves: #57008
Releases: 6.2
Change-Id: Ic595dc97826cd94d6f173b7c681194ed754ed3e0
Reviewed-on: https://review.typo3.org/28482
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Markus Klein
Tested-by: Markus Klein