[BUGFIX] Support for static table titles in new record wizard
With a static table name the extension key is not updated which sets a
wrong table title for the extension before. This patch resolves table
title from configuration with a fallback to table name part and sets
group name and extension key. This helps to prevent any conflict between
different extensions.

Resolves: #23478
Releases: 6.2
Change-Id: I5f33b0a9ab616256f0b328a6d7c793677213531b
Reviewed-on: https://review.typo3.org/28510
Reviewed-by: Vladimir Podkovanov
Reviewed-by: Wouter Wolters
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind