[BUGFIX] Own Layout Breaks Mail Message
With this patch it is possible to configure custom layouts
for the form, confirmation, and each type of post processor.
The backward compatibility is given, because this does not override
or break already existing configurations but provides some more
options to configure specific steps of the form process independent
of each other.
Therefore it prevents from breaking the layout of the confirmation view
and the HTML mail.

Resolves: #31951
Releases: 6.2
Change-Id: I3c015b53cd6d6efbd070d16bdc2111c7d60671aa
Reviewed-on: https://review.typo3.org/28526
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Reviewed-by: Markus Klein
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov