[BUGFIX] ClassLoader exits early for invalid class names
Instead of doing calculations if an invalid
class name could still be found and implementing
safeguards that our caching layer does not
fail with invalid identifiers, we exit early
if we find an invalid class name.

Maybe PHP will be smart enough at some point
not calling an autoloader at all if the class name
is invalid anyway so we could the remove this check.

By doing a comprehensive check upfront, we can
remove the safeguard code for the caching layer.

The performance impact of this change breaks down
to be lower than 1ms for a regular TYPO3 request.

Resolves: #57183
Releases: 6.2
Change-Id: Id38bea6b85fbfd280417adca325411410cd9fb63
Reviewed-on: https://review.typo3.org/28553
Reviewed-by: Thomas Maroschik
Reviewed-by: Sebastian Fischer
Reviewed-by: Alexander Stehlik
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Ernesto Baschny
Tested-by: Ernesto Baschny