[BUGFIX] First install lists directory errors
For a first installer the directory structure is empty and the
first step will try to create it. We do not need to show these
as errors to the user, only if the root directory cannot be
written.

To achieve this we raise "Warnings" instead (which are
not shown in first step). To keep the Folder Structure menu
item on the left menu in the Install Tool showing a red badge
for non existing directories, we change this to also include
Warnings.

Resolves: #57152
Releases: 6.2
Change-Id: Idd628ecdf946f9befdb7fefd7a9bb7bccc175c2b
Reviewed-on: https://review.typo3.org/28605
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Wouter Wolters
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe