[BUGFIX] Module Menu throws PHP warning for top level menu items
When rendering the module menu icons, the menu items
on the top level (Web, File, User, Admin) don't have an icon,
however the getModuleIcon functionality processes them,
and checks for getimagesize() even though no image is added.

To overcome this, the function needs to be modified slightly to
only call getimagesize if there is actually an icon.

Releases: 6.2, 6.1
Resolves: #57179
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Stefan Neufeind
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Change-Id: I1fdede253c00df9c36bb7063edf0b5bf320bb20e
Reviewed-on: https://review.typo3.org/28668
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack