[BUGFIX] Unresolvable dependency if composer.json lacks version
This change fixes a problem which occurs under these circumstances:

* Extension A depends on Extension B
* Extension B is currently installed
* Extension A is not currently installed
* Extension B has composer.json file
* Extension B composer.json file contains no "version" attribute
  or this attribute is empty.

When installing Extension A in this case, EM will fail with an
error saying "could not resolve dependency for xyz" which is a
bit too vague to be helpful. This change fixes this murky error
by throwing a much more specific error saying exactly that the
error stems from the composer.json file missing a "version" no.

Releases: 6.2
Resolves: #57216
Change-Id: I39ab18d6cf35b69600663524827422c732b813cf
Reviewed-on: https://review.typo3.org/28683
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Markus Klein
Tested-by: Markus Klein