[BUGFIX] Backup of LocalConfiguration.php missing on update
If you update from a 6.x installation there should be a backup of the
old, unmodified LocalConfiguration.php as it's getting changed during
PackageStates.php extraction. This is prevented with commit e755bcd
as a check for the PackageStates.php file was introduced, which might
not be available at this point during upgrade.

Resolves: #57458
Releases: 6.2
Change-Id: Ic3634ab58e3e4112df5d8544cca9c42e8f1db91f
Reviewed-on: https://review.typo3.org/28980
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind