[BUGFIX] Remove the need of PackageStates.php in tests
Although there is code in the package manager tests
to mock a PackageStates.php file, it does not work
as the path to the file is hardcoded in the initialize
action.

Move the code for setting the path and creating
the package factory to the constructor
and move the mocking of the path before initializing
in the test.

Resolves: #57490
Releases: 6.2
Change-Id: I2ca29016de288017ec73ff36d0dbed1e2be0e67d
Reviewed-on: https://review.typo3.org/29021
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Thomas Maroschik
Reviewed-by: Stefan Neufeind
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn