[BUGFIX] Add l18n_parent to tt_content TCA
If an editor ignores the "Copy default content elements"-button, it is
not possible to get a new content element in the alternative language
related to it's parent.

The existing l18n_parent field is never shown. This seems to be a bug
which was introduced in TYPO3 4.5 and exists up to now.

The patch adds the l18n_parent field to the general tab.

Resolves: #26709
Releases: 6.2
Change-Id: Id5f171cd4d1a6441cb58123aaa2b9bd337271556
Reviewed-on: https://review.typo3.org/29054
Reviewed-by: Frans Saris
Reviewed-by: Nadir S.
Tested-by: Nadir S.
Reviewed-by: Sascha Wilking
Tested-by: Sascha Wilking
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader